+++ jdaviz/configs/imviz/plugins/coords_info/coords_info.py
@@ -389,13 +389,10 @@ class CoordsInfo(TemplateMixin, DatasetSelectMixin):
 
     def _spectrum_viewer_update(self, viewer, x, y):
         def _cursor_fallback():
-            statistic = getattr(viewer.state, 'function', None)
-            cache_key = (viewer.state.layers[0].layer.label, statistic)
-            sp = self.app._get_object_cache.get(cache_key, viewer.data()[0])
             self._dict['axes_x'] = x
-            self._dict['axes_x:unit'] = sp.spectral_axis.unit.to_string()
+            self._dict['axes_x:unit'] = viewer.state.x_display_unit
             self._dict['axes_y'] = y
-            self._dict['axes_y:unit'] = sp.flux.unit.to_string()
+            self._dict['axes_y:unit'] = viewer.state.y_display_unit
             self._dict['data_label'] = ''
 
         def _copy_axes_to_spectral():
@@ -418,8 +415,6 @@ class CoordsInfo(TemplateMixin, DatasetSelectMixin):
             self.row3_text = ''
             self.icon = 'mdi-cursor-default'
             self.marks[viewer._reference_id].visible = False
-            # get the units from the first layer
-            # TODO: replace with display units once implemented
             _cursor_fallback()
             _copy_axes_to_spectral()
             return
@@ -462,17 +457,22 @@ class CoordsInfo(TemplateMixin, DatasetSelectMixin):
                                                        subset_to_apply=subset_label)
                     self.app._get_object_cache[cache_key] = sp
 
+                # Calculations have to happen in the frame of viewer display units.
+                disp_wave = sp.spectral_axis.to_value(viewer.state.x_display_unit, u.spectral())
+                disp_flux = sp.flux.to_value(viewer.state.y_display_unit,
+                                             u.spectral_density(sp.spectral_axis))
+
                 # Out of range in spectral axis.
                 if (self.dataset.selected != lyr.layer.label and
-                        (x < sp.spectral_axis.value.min() or x > sp.spectral_axis.value.max())):
+                        (x < disp_wave.min() or x > disp_wave.max())):
                     continue
 
-                cur_i = np.argmin(abs(sp.spectral_axis.value - x))
-                cur_wave = sp.spectral_axis[cur_i]
-                cur_flux = sp.flux[cur_i]
+                cur_i = np.argmin(abs(disp_wave - x))
+                cur_wave = disp_wave[cur_i]
+                cur_flux = disp_flux[cur_i]
 
-                dx = cur_wave.value - x
-                dy = cur_flux.value - y
+                dx = cur_wave - x
+                dy = cur_flux - y
                 cur_distance = math.sqrt(dx * dx + dy * dy)
                 if (closest_distance is None) or (cur_distance < closest_distance):
                     closest_distance = cur_distance
@@ -497,27 +497,34 @@ class CoordsInfo(TemplateMixin, DatasetSelectMixin):
             return
 
         self.row2_title = 'Wave'
-        self.row2_text = f'{closest_wave.value:10.5e} {closest_wave.unit.to_string()}'
-        self._dict['axes_x'] = closest_wave.value
-        self._dict['axes_x:unit'] = closest_wave.unit.to_string()
-        if closest_wave.unit != u.pix:
+        self.row2_text = f'{closest_wave:10.5e} {viewer.state.x_display_unit}'
+        self._dict['axes_x'] = closest_wave
+        self._dict['axes_x:unit'] = viewer.state.x_display_unit
+        if viewer.state.x_display_unit != u.pix:
             self.row2_text += f' ({int(closest_i)} pix)'
             if self.app.config == 'cubeviz':
                 # float to be compatible with nan
                 self._dict['slice'] = float(closest_i)
-                self._dict['spectral_axis'] = closest_wave.value
-                self._dict['spectral_axis:unit'] = closest_wave.unit.to_string()
+                self._dict['spectral_axis'] = closest_wave
+                self._dict['spectral_axis:unit'] = viewer.state.x_display_unit
             else:
                 # float to be compatible with nan
                 self._dict['index'] = float(closest_i)
 
+        if viewer.state.y_display_unit is None:
+            flux_unit = ""
+        else:
+            flux_unit = viewer.state.y_display_unit
         self.row3_title = 'Flux'
-        self.row3_text = f'{closest_flux.value:10.5e} {closest_flux.unit.to_string()}'
-        self._dict['axes_y'] = closest_flux.value
-        self._dict['axes_y:unit'] = closest_flux.unit.to_string()
+        self.row3_text = f'{closest_flux:10.5e} {flux_unit}'
+        self._dict['axes_y'] = closest_flux
+        self._dict['axes_y:unit'] = viewer.state.y_display_unit
 
-        self.icon = closest_icon
+        if closest_icon is not None:
+            self.icon = closest_icon
+        else:
+            self.icon = ""
 
-        self.marks[viewer._reference_id].update_xy([closest_wave.value], [closest_flux.value])  # noqa
+        self.marks[viewer._reference_id].update_xy([closest_wave], [closest_flux])
         self.marks[viewer._reference_id].visible = True
         _copy_axes_to_spectral()