+++ CHANGES.rst
@@ -25,6 +25,8 @@ Mosviz
 Specviz
 ^^^^^^^
 
+* Re-enabled unit conversion support. [#2048]
+
 Specviz2d
 ^^^^^^^^^